requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39hb82d6ee_0
aiosignal=1.2.0=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
async-timeout=4.0.1=pyhd8ed1ab_0
attrs=21.2.0=pyhd8ed1ab_0
audioread=2.1.9=py39hcbf5309_1
blas=2.112=mkl
blas-devel=3.9.0=12_win64_mkl
blinker=1.4=py_1
brotli=1.0.9=h8ffe710_6
brotli-bin=1.0.9=h8ffe710_6
brotlipy=0.7.0=py39hb82d6ee_1003
ca-certificates=2021.10.8=h5b45459_0
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2021.10.8=py39hcbf5309_1
cffi=1.15.0=py39h0878f49_0
charset-normalizer=2.0.8=pyhd8ed1ab_0
click=8.0.3=py39hcbf5309_1
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py39h7bc7c5c_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
decorator=5.1.0=pyhd8ed1ab_0
dill=0.3.4=pyhd8ed1ab_0
ffmpeg=4.3.1=ha925a31_0
fonttools=4.28.2=py39hb82d6ee_0
freetype=2.10.4=h546665d_1
frozenlist=1.2.0=py39hb82d6ee_1
google-auth=2.3.3=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py39hb76b349_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.8.2=py39hcbf5309_0
intel-openmp=2021.4.0=h57928b3_3556
jbig=2.1=h8d14728_2003
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
kiwisolver=1.3.2=py39h2e07f2f_1
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_6
libbrotlidec=1.0.9=h8ffe710_6
libbrotlienc=1.0.9=h8ffe710_6
libcblas=3.9.0=12_win64_mkl
libdeflate=1.8=h8ffe710_0
libflac=1.3.3=h0e60522_1
liblapack=3.9.0=12_win64_mkl
liblapacke=3.9.0=12_win64_mkl
libogg=1.3.4=h8ffe710_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.37=h1d00b33_2
libprotobuf=3.19.1=h7755175_0
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.31=h0e60522_1
libtiff=4.3.0=hd413186_2
libuv=1.42.0=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
libzlib=1.2.11=h8ffe710_1013
llvmlite=0.37.0=py39ha0cd8c8_1
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.6=pyhd8ed1ab_0
matplotlib-base=3.5.0=py39h581301d_0
mkl=2021.4.0=h0e2418a_729
mkl-devel=2021.4.0=h57928b3_730
mkl-include=2021.4.0=h0e2418a_729
msys2-conda-epoch=20160418=1
multidict=5.2.0=py39hb82d6ee_1
multiprocess=0.70.12.2=py39hb82d6ee_1
munkres=1.1.4=pyh9f0ad1d_0
numba=0.54.1=py39hb8cd55e_0
numpy=1.20.3=py39h6635163_1
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb211442_1
openssl=1.1.1l=h8ffe710_0
p-tqdm=1.3.3=pyhd8ed1ab_0
packaging=21.3=pyhd8ed1ab_0
pathos=0.2.8=pyhd8ed1ab_0
pillow=8.4.0=py39h916092e_0
pip=21.3.1=pyhd8ed1ab_0
pooch=1.5.2=pyhd8ed1ab_0
pox=0.3.0=pyhd8ed1ab_0
ppft=1.6.6.4=pyhd8ed1ab_0
protobuf=3.19.1=py39h415ef7b_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=1.7.1=py_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.6=pyhd8ed1ab_0
pysocks=1.7.1=py39hcbf5309_4
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.9.7=h7840368_3_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.0=py3.9_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pyu2f=0.1.5=pyhd8ed1ab_0
requests=2.26.0=pyhd8ed1ab_1
requests-oauthlib=1.3.0=pyh9f0ad1d_0
resampy=0.2.2=py_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.0.1=py39he931e04_2
scipy=1.7.3=py39hc0c34ad_0
setuptools=59.4.0=py39hcbf5309_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.37.0=h8ffe710_0
tbb=2021.4.0=h2d74725_1
tensorboard=2.7.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py39hcbf5309_1
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
threadpoolctl=3.0.0=pyh8a188c0_0
tk=8.6.11=h8ffe710_1
torchaudio=0.10.0=py39_cu113
torchvision=0.11.1=py39_cu113
tqdm=4.62.3=pyhd8ed1ab_0
typing-extensions=4.0.1=hd8ed1ab_0
typing_extensions=4.0.1=pyha770c72_0
tzdata=2021e=he74cb21_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.7=pyhd8ed1ab_0
vc=14.2=hb210afc_5
vs2015_runtime=14.29.30037=h902a5da_5
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_1
win_inet_pton=1.1.0=py39hcbf5309_3
xz=5.2.5=h62dcd97_1
yarl=1.7.2=py39hb82d6ee_1
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h8ffe710_1013
zstd=1.5.0=h6255e5f_0
